@@ -3,8 +3,8 @@
     "type": "phpbb-extension",
     "description": "Adds the option to disable and re-enable all installed extensions at once to the extension management. Individual selections can be made using checkboxes. In addition, sequence groups and dependencies can be defined or extensions can be ignored. The version check is carried out in blocks and provides detailed information. Additional information is displayed in the extension management.",
     "homepage": "https://github.com/LukeWCS/ext-mgr-plus",
-    "version": "3.1.0-b17",
-    "time": "2025-06-02",
+    "version": "3.1.0-b18",
+    "time": "2025-07-14",
     "license": "GPL-2.0-only",
     "authors": [
         {
@@ -19,7 +19,7 @@
         }
     ],
     "require": {
-        "php": ">=7.4.0,<8.5.0@dev",
+        "php": ">=8.0.0,<8.5.0@dev",
         "composer/installers": "~1.0.0"
     },
     "extra": {
